+++ rosen/modules/render_service_base/include/animation/rs_render_animation.h
@@ -60,7 +60,7 @@ public:
     void DumpAnimation(std::string& out) const;
     virtual void DumpAnimationType(std::string& out) const;
 
-    virtual uint16_t GetRemainingTime(const bool isCountRepeat = false) const
+    virtual int64_t GetRemainingTime(const bool isCountRepeat = false) const
     {
         return 0;
     }
